+++ server/internal/app/grpc.go
@@ -172,7 +172,7 @@ func isReadOnlyMethod(method string) bool {
 		"v1.ReEarthVisualizer/GetAllProjects",
 		"v1.ReEarthVisualizer/GetProjectList",
 		"v1.ReEarthVisualizer/GetProject",
-		"v1.ReEarthVisualizer/GetProjectByAlias",
+		"v1.ReEarthVisualizer/GetProjectByWorkspaceAliasAndProjectAlias",
 		"v1.ReEarthVisualizer/ExportProject",
 	}