+++ web/src/app/lib/reearth-widget-ui/icons/index.ts
@@ -1,0 +1,7 @@
+export * from "./Eye";
+export * from "./EyeSlash";
+export * from "./Play";
+export * from "./Repeat";
+export * from "./RepeatOnce";
+export * from "./SkipBack";
+export * from "./SkipForward";