+++ web/src/ee/featureConfig.ts
@@ -11,6 +11,7 @@ export const getFeatureConfig = (): AppFeatureConfig => {
     externalWorkspaceManagementUrl: `${c?.platformUrl}/[WORKSPACE_ALIAS]/settings`,
     accountManagement: false,
     projectVisibility: true,
-    externalAccountManagementUrl: `${c?.platformUrl}/settings/profile`
+    externalAccountManagementUrl: `${c?.platformUrl}/settings/profile`,
+    builtinTimelineWidget: false
   };
 };