+++ web/src/services/api/user/index.ts
@@ -1,4 +1,3 @@
 export * from "./useMeQueries";
 export * from "./useMeMutations";
 export * from "./useUserQueries";
-export * from "./useUserMutations";